fix: fix commit message linting
Before this fix, the commit message linting was linting your previous commit message,
not your current one. If you had a fresh clone of the repo - and therefore didn't have
a previous commit message - then the linting would fail and you were not able to commit.

The cause of this issue is that the linting was running in the precommit hook, which runs
before a commit message for the current commit is written. Therefore we did not pass an
argument to "commitlint --edit" and it defaulted to looking at your last commit message
(or failing if that does not exist).

The fix was the run the commit message linting in the commit-msg hook, which runs after a
commit message is written for the current commit, but before the commit is finalised.